refactor:ヘッダーをロジックとコンポーネントに分離
